split font code into own file
